Minor changes & fiddles

+++ R/GetUnzip.R
@@ -1,8 +1,13 @@
 # Function to extract csv data from a zipped archive
+#' @export GetUnzip
 GetUnzip=function(ZipName, FileName) {
-  temp <- tempfile()
-  download.file(ZipName,temp)
-  data <- read.csv(unz(temp, FileName))
-  unlink(temp)
+  if(grepl('^[hf]t+p', ZipName)) {
+    temp <- tempfile()
+    download.file(ZipName,temp)
+    data <- read.csv(unz(temp, FileName))
+    unlink(temp)
+  } else {
+    data <- read.csv(unz(ZipName, FileName))
+  }
   data
 }
